Update so that VPN is enabled only on the primary connection
Before the update, the vpn would only be enabled on the first found connection without checking if it has default routes or not (it could even be bridges). Normally if a connection has default routes then it means that is has connection to some sort of a network, even if it does not have internet connection.

Now the vpn is enabled on a connection that has default routes (either only IPv4 or IPv4 and IPv6), thus minizing the risk of connecting the vpn to a non-functional connection.