Custom error handling
New exceptions were created to know exactly why server issues might occur.

EmptyServerListError is the generic exceptions that is thrown when the filter returns zero servers. Although first we need to know if we have an existent cache, if not then a specific exception must be thrown for non-existent cache.