Add lib version to service file
By adding the version at the top, the systemd service will be recreated if the hashes do not match. The hashes will no match if they have different library versions, thus a force restart will be made everytime the package will be upgraded, thus re-assuring that the service is working properly.